LPS-70805 Deal with integer and double numbers

+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-expression/src/main/java/com/liferay/dynamic/data/mapping/expression/internal/DDMExpressionImpl.java
@@ -153,12 +153,12 @@ public class DDMExpressionImpl<T> implements DDMExpression<T> {
 	public void setIntegerVariableValue(
 		String variableName, Integer variableValue) {
 
-		setVariableValue(variableName, variableValue.doubleValue());
+		setVariableValue(variableName, variableValue.intValue());
 	}
 
 	@Override
 	public void setLongVariableValue(String variableName, Long variableValue) {
-		setVariableValue(variableName, variableValue.doubleValue());
+		setVariableValue(variableName, variableValue.longValue());
 	}
 
 	/**
